refactor: lens-debate — readAllFontSizes duplicates readAllCols (concept multiplication)
Extracted a shared readPerTerminal projector into support/buffer.ts and reimplemented readAllCols ('cols') and readAllFontSizes ('fontSize') on top of it, deleting the duplicated enumeration/filter logic.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-1, raised by hickey). Not pushed or merged.
