+++ packages/tests/support/buffer.ts
@@ -30,6 +30,43 @@ export function readBufferText(
   );
 }
 
+/** Serializable discriminant for {@link readPerTerminal}: which per-terminal
+ *  scalar to project off each inner xterm container. */
+type PerTerminalProbe = "cols" | "fontSize";
+
+/**
+ * Enumerate the inner terminal containers — those carrying both
+ * `data-terminal-id` and `data-font-size` — and project one per-terminal
+ * scalar off each, keyed by id. The outer CanvasTile wrapper carries
+ * `data-terminal-id` but no `data-font-size`, so it's filtered out (it also
+ * never holds the `__xterm` ref). `cols` reads the live xterm grid width via
+ * the `__xterm` ref attached in Terminal.tsx's onMount; `fontSize` parses the
+ * `data-font-size` attribute. The probe is a serializable discriminant rather
+ * than a live closure because the body runs inside `page.evaluate`.
+ */
+export function readPerTerminal(
+  page: Page,
+  probe: PerTerminalProbe,
+): Promise<Record<string, number>> {
+  return page.evaluate((p) => {
+    const out: Record<string, number> = {};
+    for (const n of document.querySelectorAll(
+      "[data-terminal-id][data-font-size]",
+    )) {
+      const id = n.getAttribute("data-terminal-id");
+      if (!id) continue;
+      if (p === "cols") {
+        const term = (n as unknown as { __xterm?: { cols: number } }).__xterm;
+        if (term && typeof term.cols === "number") out[id] = term.cols;
+      } else {
+        const fs = n.getAttribute("data-font-size");
+        if (fs) out[id] = Number.parseFloat(fs);
+      }
+    }
+    return out;
+  }, probe);
+}
+
 /**
  * Wait for the xterm buffer to contain the expected text using Playwright's
  * native waitForFunction (rAF-based polling inside the browser context).
